+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-08-11 10:19:29
-// Commit: feat(state): integrate offline caching for state state
+// Updated at: 2026-08-11 13:40:28
+// Commit: perf(state): optimize memory consumption in state list viewport
 
 class stateComponentState {
-  final String commitRef = '3516';
-  final String lastMessage = 'feat(state): integrate offline caching for state state';
-  final DateTime timestamp = DateTime.parse('2026-08-11T10:19:29+05:00');
+  final String commitRef = '3524';
+  final String lastMessage = 'perf(state): optimize memory consumption in state list viewport';
+  final DateTime timestamp = DateTime.parse('2026-08-11T13:40:28+05:00');
 }